backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Email
sib-api-v3-sdk==7.6.0  # Brevo API (PRIMARY - formerly Sendinblue)
mailjet-rest==1.3.4  # Mailjet API (STANDBY)
requests==2.31.0  # For MailBluster API

# QR Code generation
qrcode[pil]==7.4.2
Pillow==11.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
